requirements.txt
python-dotenv==1.0.0
discord.py==2.3.2
openai==1.3.0
anthropic==0.7.8
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic==2.5.0
aiofiles==23.2.1